fix promotion_history

+++ database/migrations/2025_07_11_154714_create_author_book_table.php
@@ -8,17 +8,19 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('author_book', function (Blueprint $table) {
-            $table->string('author_id', 50);
-            $table->string('book_id', 50);
+        if (!Schema::hasTable('author_book')) {
+            Schema::create('author_book', function (Blueprint $table) {
+                $table->string('author_id', 50);
+                $table->string('book_id', 50);
 
-            $table->primary(['author_id', 'book_id']);
+                $table->primary(['author_id', 'book_id']);
 
-            $table->foreign('author_id')->references('id')->on('authors')->onDelete('cascade');
-            $table->foreign('book_id')->references('id')->on('books')->onDelete('cascade');
+                $table->foreign('author_id')->references('id')->on('authors')->onDelete('cascade');
+                $table->foreign('book_id')->references('id')->on('books')->onDelete('cascade');
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
     }
 
     public function down(): void
